# src/swagger.yaml
openapi: 3.0.4
info:
  title: IMLGS Mirror
  description: |-
    This is a mirror of the Index to Marine and Lacustrine Geological Samples (IMLGS). The orignal was at 
      [https://www.ncei.noaa.gov/products/index-marine-lacustrine-samples](https://www.ncei.noaa.gov/products/index-marine-lacustrine-samples) and data access at [https://www.ncei.noaa.gov/maps/imlgs/samples](https://www.ncei.noaa.gov/maps/imlgs/samples). 
   
  version: 1.0.0
servers:
  - url: https://localhost:3000/api
...
paths:
  /cruises/name:
    get:
      tags:
        - cruises
      summary: Get cruise names 
      description: Get minimal cruise information
      operationId: cruiseName
      parameters:
        - name: year
          in: query
          description: Cruise year
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CruiseNamePage'
  /cruises/detail:
    get:
      tags:
        - cruises
      summary: Get cruise details 
      description: Get full cruise information
      operationId: cruiseDetail
      parameters:
        - name: year
          in: query
          description: Cruise year
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CruisePage'
  /cruises/detail/{id}:
    get:
      tags:
        - cruises
      summary: Get cruise details 
      description: Get full cruise information
      operationId: cruiseDetailId
      parameters:
        - name: id
          in: path
          description: Cruise id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cruise'
...
components:
  schemas:
    CruiseName:
      type: object
      properties:
        id:
          type: integer
          format: int64 
        cruise:
          type: string
        year:
          type: integer
        legs:
          type: array
          items: 
            type: string
    PlatformName:
      type: object
      properties:
        id: 
          type: integer
        platform:
          type: string
    Cruise:
      type: object
      properties:
        id:
          type: integer
          format: int64 
        cruise:
          type: string
        year:
          type: integer
        legs:
          type: array
          items: 
            type: string
        facilities:
          type: string
        platforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformName"
    CruiseNamePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CruiseName"
        page: 
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        items_per_page: 
          type: integer
    CruisePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Cruise"
        page: 
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        items_per_page: 
          type: integer
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            "write:pets": modify pets in your account
            "read:pets": read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header